Assignment 3 is about linear regression
In this code we are using gradient descent for linear regression. This is an optional method to be used in linear regression. We can also solve the linear regression using matrix multiplication which minimizws the error. Linear regression is one of the most used algorithm for fitting the graph in a line in the field of machine learning.